+++ launcher/src/com/genir/aitweaks/launcher/Entrypoints.kt
@@ -16,4 +16,6 @@ interface ShipAIPicker {
     fun pickShipAI(member: FleetMemberAPI?, ship: ShipAPI): PluginPick<ShipAIPlugin>?
 
     fun canHaveCustomAI(ship: ShipAPI): Boolean
+
+    fun getUnapplicableReason(ship: ShipAPI): String?
 }