+++ changelog.md
@@ -1,5 +1,6 @@
 - Custom AI: Fixed an issue where backing off ships would collide with enemies.
 - DEPRECATED the Search and Destroy hullmod. Now ships with Custom AI hullmod will default to Search and Destroy order.
+- Added descriptions explaining why hullmods can not be installed on given ships.
 
 Version v2.2.0